app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
uvicorn = "^0.11.3"
fastapi = "^0.54.1"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.23.0"
celery = "^4.4.2"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
tenacity = "^6.1.0"
pydantic = "^1.9"
emails = "^0.5.15"
raven = "^6.10.0"
gunicorn = "^20.0.4"
jinja2 = "^2.11.2"
psycopg2-binary = "^2.8.5"
alembic = "^1.4.2"
sqlalchemy = "^1.3.16"
pytest = "^5.4.1"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
boto3 = "<1.23.2"
uuid = "^1.30"
pandas = "^1.4.2"
humps = "^0.2.2"
MarkupSafe = "2.0.1"
torch = "1.9.0"
torch-geometric = "^2.0.4"
rdkit-pypi = "^2022.3.2"
torch-sparse = "0.6.12"
torch-scatter = "^2.0.9"
PyYAML = "^6.0"
torchsummary = "^1.5.1"
pytest-dotenv = "^0.5.2"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^21.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 80
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

